+++ app/admin/lists/ListBuilderPage.tsx
@@ -21,6 +21,7 @@ import { AppearancePanel } from './AppearancePanel';
 import { FieldBuilder } from './FieldBuilder';
 import { MaintenancePanel } from './MaintenancePanel';
 import { PermissionsPanel } from './PermissionsPanel';
+import { PortalConfigPanel } from './PortalConfigPanel';
 import { PublicVisibilityPanel } from './PublicVisibilityPanel';
 import { SlugEditor } from './SlugEditor';
 
@@ -215,6 +216,7 @@ export function ListBuilderPage(): JSX.Element {
             <AppearancePanel list={list.data} />
             <PermissionsPanel listId={list.data.id} />
             <PublicVisibilityPanel list={list.data} />
+            <PortalConfigPanel list={list.data} />
             <MaintenancePanel listId={list.data.id} />
         </div>
     );